Invert the mailbox pump so the kernel drains and gameplay applies (#368)
`process_represented_session_commands_like_cpp` drained the two rails and then
named, arm by arm, the gameplay each `SessionCommand` performs. 31 of the 38
methods `mailbox/` called on `self` were defined in `handlers/`, which is the
largest single reason #297's kernel cannot be a crate: a kernel containing the
mailbox would depend on gameplay.

The 37 arms move to `crate::session_commands` unchanged. The pump keeps what is
queue mechanics — rail ordering, the durable-overflow kick, backpressure — and
hands each command to one entry point. `mailbox/` now names 3 methods outside
itself instead of 38, and 1 in `handlers/` instead of 31. Across the six
candidate kernel modules the union falls from 83 to 49.

## Why the arms moved intact rather than becoming thunks

This looks like the opcode table #359 retired and is not. That match is over an
enum with 37 arms and no wildcard, so the compiler already enforces the pairing:
a new variant that nobody applies does not build. There is no second declaration
and no silent-drop failure mode. Giving each command a thunk would have bought a
symmetrical-looking change and thrown away the exhaustiveness — a worse boundary
than the one it replaced. Only the dependency direction was wrong, so only the
dependency direction changed.

## What is pinned

Two tests cover what the kernel still owns alone, which nothing exercised before:
that draining presents the durable prefix, then the general rail, then the
deferred durable suffix; and that an overflowed authoritative backlog kicks the
session *before* any command is applied, leaving the queued command untouched.

## The call that stays

`apply_pending_durable_item_loot_completions_like_cpp` is still named by the
pump. It is a four-line loot wrapper that must run before the overflow check so
a session about to be disconnected still lands committed loot. Wrapping it in a
second one-implementation indirection would reach zero on the count and leave a
worse boundary, which the epic's principle 8 forbids; the honest seam is the
loot rail draining the way the command rails now do, and that is loot's work.

## Ledger

`session/mod.rs` production falls 73,057 -> 72,936. The syntax baseline gains
one impl owner and one associated item, both `crate::session_commands`. That
registers as external impl growth against the hotspot's retirement condition
even though it is code leaving `session/`, not new surface — recorded here
rather than left to look like an oversight. The row also drops #359 from its
open retirement issues, and the mirrored issue state is refreshed to match.

Checks: `cargo test -p wow-world --lib` 3573/0, `-p world-server` 449/0;
`check_architecture.py check`/`self-test` PASS; `session-ownership-check`
syntax-only and exhaustive PASS; `./tools/validation-v2 final` PASS.

Closes #368.
