+++ crates/wow-world/src/lib.rs
@@ -17,6 +17,7 @@ pub mod phasing;
 pub(crate) mod profession;
 pub mod reputation;
 pub mod session;
+mod session_commands;
 mod session_policy;
 #[allow(dead_code)] // Private prerequisite seam consumed by trainer issue #157.
 pub(crate) mod spell_acquisition;
