+++ crates/wow-world/src/session_tests.rs
@@ -23,6 +23,8 @@ mod driver;
 mod lifecycle;
 #[path = "session/tests/lifecycle_persistence.rs"]
 mod lifecycle_persistence;
+#[path = "session/tests/mailbox_pump.rs"]
+mod mailbox_pump;
 #[path = "session/tests/routing.rs"]
 mod routing;
 #[path = "session/tests/save_plan_order.rs"]
