+++ tools/architecture/runtime-ownership-ledger.json
@@ -1820,16 +1820,15 @@
       "entries": [
         {
           "path": "crates/wow-world/src/session/mod.rs",
-          "production_lines": 73057,
-          "test_lines": 95801,
-          "total_lines": 168858,
+          "production_lines": 72936,
+          "test_lines": 95939,
+          "total_lines": 168875,
           "owner": "wow-world WorldSession/application monolith.",
           "open_retirement_issues": [
             153,
-            297,
-            359
+            297
           ],
-          "retirement_condition": "Session is a thin protocol/lifecycle shell with no mutable Player/Map gameplay, concrete DB/catalog bag, or external impl growth; LOC alone does not close it. #297 owns the kernel promotion and #359 the dispatcher decoupling that unblocks it; #252 owns the gameplay mirror. Until they land this row is a reviewed exception to the epic's 4,000-line signal, not a target."
+          "retirement_condition": "Session is a thin protocol/lifecycle shell with no mutable Player/Map gameplay, concrete DB/catalog bag, or external impl growth; LOC alone does not close it. #359 landed the dispatcher decoupling and #368 the mailbox inversion; #297 owns the kernel promotion those unblock, and #252 the gameplay mirror. Until they land this row is a reviewed exception to the epic's 4,000-line signal, not a target."
         },
         {
           "path": "crates/wow-map/src/map/mod.rs",
